@@ -3,6 +3,7 @@
     using System;
     using System.Threading;
     using System.Threading.Tasks;
+using Application.Categories.Dtos;
     using AutoMapper;
     using AutoMapper.QueryableExtensions;
     using Domain.CMS;
@@ -13,8 +14,8 @@
 
     public class CategoryDetails
     {
-        public record Query(int Id) : IRequest<ActionResult<Category>>;
-        public class Handler : IRequestHandler<Query, ActionResult<Category>>
+        public record Query(int Id) : IRequest<CategoryDto>;
+        public class Handler : IRequestHandler<Query, CategoryDto>
         {
             private readonly DataContext _context;
             private readonly IMapper _mapper;
@@ -25,17 +26,12 @@
                 _mapper = mapper;
             }
 
-            public async Task<ActionResult<Category>> Handle(Query request, CancellationToken cancellationToken)
+            public async Task<CategoryDto> Handle(Query request, CancellationToken cancellationToken)
             {
-                var entity = await _context.Categories.ProjectTo<Category>(_mapper.ConfigurationProvider)
+                return await _context.Categories
+                    .AsNoTracking()
+                    .ProjectTo<CategoryDto>(_mapper.ConfigurationProvider)
                     .FirstOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
-
-                if (entity == null)
-                {
-                    return new NotFoundResult();
-                }
-
-                return entity;
             }
         }
 
