@@ -3,6 +3,7 @@
     using System;
     using System.Threading;
     using System.Threading.Tasks;
+using Application.Menus.Dtos;
     using AutoMapper;
     using AutoMapper.QueryableExtensions;
     using Domain.CMS;
@@ -13,8 +14,8 @@
         
     public class MenuDetails
     {
-        public record Query(int Id) : IRequest<ActionResult<Menu>>;
-        public class Handler : IRequestHandler<Query, ActionResult<Menu>>
+        public record Query(int Id) : IRequest<MenuDto>;
+        public class Handler : IRequestHandler<Query, MenuDto>
         {
             private readonly DataContext _context;
             private readonly IMapper _mapper;
@@ -25,17 +26,12 @@
                 _mapper = mapper;
             }
 
-            public async Task<ActionResult<Menu>> Handle(Query request, CancellationToken cancellationToken)
+            public async Task<MenuDto> Handle(Query request, CancellationToken cancellationToken)
             {
-                var menu = await _context.Menus.ProjectTo<Menu>(_mapper.ConfigurationProvider)
+                return await _context.Menus
+                    .AsNoTracking()
+                    .ProjectTo<MenuDto>(_mapper.ConfigurationProvider)
                     .FirstOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
-
-                if (menu == null)
-                {
-                    return new NotFoundResult();
-                }
-
-                return menu;
             }
         }
 
