@@ -8,32 +8,31 @@ namespace Application.Posts.Dtos
 {
     using System;
     using System.Collections.Generic;
+    using Application.Categories.Dtos;
     using Domain.CMS;
     public record PostDetailsDto
     {
-        public int Id { get; set; }
+        public int Id { get; init; }
 
-        public int CategoryId { get; set; }
-        public int DomainId { get; set; }
-        public bool IsFeatured { get; set; }
+        public int CategoryId { get; init; }
+        public bool IsFeatured { get; init; }
 
-        public List<string> Files { get; set; }
+        public List<string> Files { get; init; }
 
-        public List<string> Images { get; set; }
+        public List<string> Images { get; init; }
 
-        public string? IntroText { get; set; }
+        public string? IntroText { get; init; }
 
-        public bool IsPublished { get; set; }
+        public bool IsPublished { get; init; }
 
-        public Domain Domain { get; set; }
-        public Category Category { get; set; }
+        public CategoryDto Category { get; init; }
 
-        public DateTime PublishDate { get; set; }
+        public DateTime PublishDate { get; init; }
 
-        public string Slug { get; set; }
+        public string Slug { get; init; }
 
-        public string? Text { get; set; }
+        public string? Text { get; init; }
 
-        public string Title { get; set; }
+        public string Title { get; init; }
     }
 }
