@@ -9,18 +9,14 @@
     {
         public PostProfiles()
         {
-            CreateMap<Post, Post>();
             CreateMap<Post, PostDto>();
             CreateMap<Post, PostDetailsDto>();
             CreateMap<PostCreateDto, Post>();
             CreateMap<PostEditDto, Post>();
+
             CreateMap<PostPatchDto, Post>()
                 .ForMember(x => x.IsFeatured, x => x.Condition(p => p.IsFeatured != null))
                 .ForMember(x => x.IsPublished, x => x.Condition(p => p.IsPublished != null));
-            CreateMap<Post, PublicPostDto>()
-                .ForMember(x => x.Url, x => x.MapFrom(p => $"/{p.Category.Slug}/{p.Id}/{p.Slug}"))
-                .ForMember(x => x.Language, x => x.MapFrom(p => p.Category.Language.Slug));
-            CreateMap<Post, PublicPostDetailsDto>();
         }
     }
 }
