@@ -1,23 +1,24 @@
-﻿namespace Application.Posts
-{
-    using System;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using AutoMapper;
-    using Domain.CMS;
-    using Dtos;
-    using FluentValidation;
-    using MediatR;
-    using Microsoft.AspNetCore.Http;
-    using Microsoft.AspNetCore.Mvc;
-    using Persistence;
-    using Validation;
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Application.Posts.Dtos;
+using Application.Posts.Validation;
+using AutoMapper;
+using Domain.CMS;
+using FluentValidation;
+using MediatR;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Persistence;
+
+namespace Application.Posts;
+
 
     public class PostCreate
     {
-        public record Command(PostCreateDto Post) : IRequest<ActionResult<PostDetailsDto>>;
+        public record Command(PostCreateDto Post) : IRequest<PostDetailsDto>;
 
-        public class Handler : IRequestHandler<Command, ActionResult<PostDetailsDto>>
+        public class Handler : IRequestHandler<Command, PostDetailsDto>
         {
             private readonly DataContext _context;
             private readonly IMapper _mapper;
@@ -28,15 +29,14 @@
                 _mapper = mapper;
             }
 
-            public async Task<ActionResult<PostDetailsDto>> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<PostDetailsDto> Handle(Command request, CancellationToken cancellationToken)
             {
                 var post = _mapper.Map(request.Post, new Post());
 
                 _context.Posts.Add(post);
                 await _context.SaveChangesAsync(cancellationToken);
 
-                return new OkObjectResult(_mapper.Map(post, new PostDetailsDto()))
-                    { StatusCode = StatusCodes.Status201Created };
+                return _mapper.Map(post, new PostDetailsDto()); 
             }
 
             public class CommandValidator : AbstractValidator<Command>
@@ -47,5 +47,4 @@
                 }
             }
         }
-    }
 }
