@@ -5,17 +5,20 @@
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
+    using Application.Extensions;
+    using Application.Features;
     using AutoMapper;
     using AutoMapper.QueryableExtensions;
     using Domain.CMS;
     using Dtos;
     using MediatR;
+using Microsoft.AspNetCore.Mvc.RazorPages;
     using Microsoft.EntityFrameworkCore;
     using Persistence;
 
     public class PostList
     {
-        public record Query(int PageNr) : IRequest<List<PostDto>>;
+        public record Query(PaginationDto Pagination) : IRequest<List<PostDto>>;
         public class Handler : IRequestHandler<Query, List<PostDto>>
         {
             private readonly DataContext _context;
@@ -30,16 +33,12 @@
 
             public async Task<List<PostDto>> Handle(Query request, CancellationToken cancellationToken)
             {
-                const int PostsPerPage = 10;
-                var page = Math.Max(request.PageNr - 1, 0);
-                var postsToSkip = page * PostsPerPage;
-                   
                 return await _context.Posts 
+                    .AsNoTracking()
                     .ProjectTo<PostDto>(_mapper.ConfigurationProvider)
                     .OrderByDescending(x => x.IsFeatured)
                     .ThenByDescending(x => x.PublishDate)
-                    .Skip(postsToSkip)
-                    .Take(PostsPerPage)
+                    .Paginate(request.Pagination)
                     .ToListAsync(cancellationToken);
             }
         }
