@@ -17,8 +17,6 @@
 
         public DbSet<Menu> Menus { get; init; }
 
-        public DbSet<Domain> Domains { get; init; }
-
         public DbSet<Post> Posts { get; init; }
 
         protected override void OnModelCreating(ModelBuilder builder)
@@ -29,21 +27,18 @@
             CreateCategoryModel(builder);
             CreateLanguageModel(builder);
             CreateMenuModel(builder);
-            CreateDomainModel(builder);
         }
 
         private static void CreateMenuModel(ModelBuilder builder)
         {
             builder.Entity<Menu>().HasOne(x => x.Category).WithMany();
-            builder.Entity<Menu>().HasOne(x => x.Domain).WithMany();
             builder.Entity<Menu>().HasOne(x => x.ParentMenu).WithMany().OnDelete(DeleteBehavior.Cascade);
-            builder.Entity<Menu>().HasIndex(x => new { x.Slug, x.CategoryId, x.DomainId }).IsUnique();
+            builder.Entity<Menu>().HasIndex(x => new { x.Slug, x.CategoryId }).IsUnique();
         }
 
         private static void CreatePostModel(ModelBuilder builder)
         {
             builder.Entity<Post>().HasOne(x => x.Category).WithMany();
-            builder.Entity<Post>().HasOne(x => x.Domain).WithMany();
             builder.Entity<Post>().HasIndex(x => x.Title).IsTsVectorExpressionIndex("lithuanian");
         }
 
@@ -58,9 +53,5 @@
             builder.Entity<Language>().HasIndex(x => x.Slug).IsUnique();
         }
 
-        private static void CreateDomainModel(ModelBuilder builder)
-        {
-            builder.Entity<Domain>().HasIndex(x => x.Slug).IsUnique();
-        }
     }
 }
